refactor: Collect telemetry event using Microsoft Telemetry Extension API
