+++ clients/cobol-lsp-vscode-extension/src/services/CopybookFix.ts
@@ -32,7 +32,7 @@ export class CopybookFix {
         const actionSettings = "Change settings";
         const action = await vscode.window.showErrorMessage(title, actionSettings);
         if (action === actionSettings) {
-            vscode.commands.executeCommand("workbench.action.openSettings", "broadcom-cobol-lsp");
+            vscode.commands.executeCommand("broadcom-cobol-lsp.cpy-manager.goto-settings", "broadcom-cobol-lsp");
         }
     }
