+++ clients/cobol-lsp-vscode-extension/src/services/reporter/TelemetryService.ts
@@ -11,21 +11,31 @@
  * Contributors:
  *   Broadcom, Inc. - initial API and implementation
  */
-import {ExtensionUtils} from "../settings/util/ExtensionUtils";
+import {userInfo} from "os";
+import {sep} from "path";
 import {TelemetryEvent} from "./model/TelemetryEvent";
 import {TelemetryFactory} from "./TelemetryFactory";
 
 export class TelemetryService {
+    /**
+     *  This method calculate the time elapsed between two number representation of dates expressed as
+     *  number of milliseconds elapsed since January 1, 1970 00:00:00 UTC.
+     * @param startTime numeric representation of a starting time
+     * @param endTime numeric representation of a ending time
+     */
+    public static calculateTimeElapsed(startTime: number, endTime: number): number | undefined {
+        return ((startTime && endTime) && (endTime >= startTime)) ? ((endTime - startTime) / 1000) : undefined;
+    }
 
     /**
-     * This method collect the data event for telemetring purpose, apply basic validation and delegate the reporter service to sent them to the target telemetry server
+     * This method collect the data event for telemetry purpose, apply basic validation
+     * and delegate the reporter service to sent them to the target telemetry server
      * @param eventName name assigned for the event for the telemetry
      * @param categories optional list of keywords (act as tags)
      * @param notes optional brief description
      * @param telemetryMeasurement optional set of numeric data with a key name
      */
     public static registerEvent(eventName: string, categories?: string[], notes?: string, telemetryMeasurement?: Map<string, number>) {
-        // exit if event name is empty
         if (!this.isValidEventName(eventName)) {
             return;
         }
@@ -34,9 +44,10 @@ export class TelemetryService {
     }
 
     /**
-     * This method collect the exception data event for telemetring purpose, apply basic validation and delegate the reporter service to sent them to the target telemetry server
+     * This method collect the exception data event for telemetry purpose, apply basic validation
+     * and delegate the reporter service to sent them to the target telemetry server
      * @param eventName name assigned for the event for the telemetry
-     * @param rootCause the stacktrace attached for debugging purpose. User folder is anonymized
+     * @param rootCause the stacktrace attached for debugging purpose.
      * @param categories optional list of keywords (act as tags)
      * @param notes optional brief description
      * @param telemetryMeasurement optional set of numeric data with a key name
@@ -46,7 +57,7 @@ export class TelemetryService {
             return;
         }
 
-        TelemetryFactory.getReporter().reportExceptionEvent(this.createTelemetryEvent(eventName, categories, notes, ExtensionUtils.anonymizeContent(rootCause), telemetryMeasurement));
+        TelemetryFactory.getReporter().reportExceptionEvent(this.createTelemetryEvent(eventName, categories, notes, this.anonymizeContent(rootCause), telemetryMeasurement));
     }
 
     private static isValidEventName(eventName: string): boolean {
@@ -64,12 +75,24 @@ export class TelemetryService {
         return telemetryEvent;
     }
 
+    private static resolveCategories(categories: string[]): string {
+        return (categories) ? categories.toString() : "N.D";
+    }
+
     private static isValidRootCause(rootCause: string): boolean {
         return rootCause !== undefined && rootCause.trim() !== "";
     }
 
-    private static resolveCategories(categories: string[]): string {
-        return (categories) ? categories.toString() : "N.D";
+    /**
+     * This method allows to anonymize the username value present in stack trace
+     * @param content stack trace
+     */
+    private static anonymizeContent(content: string): string {
+        return content.replace(new RegExp("\\" + sep + this.getUsername(), "g"), sep + "anonymous");
+    }
+
+    private static getUsername(): string {
+        return userInfo().username;
     }
 
 }
