+++ clients/cobol-lsp-vscode-extension/src/services/settings/util/ExtensionUtils.ts
@@ -12,8 +12,6 @@
  *   Broadcom, Inc. - initial API and implementation
  */
 
-import {userInfo} from "os";
-import {sep} from "path";
 import * as vscode from "vscode";
 import {EXTENSION_ID} from "../../../constants";
 
@@ -22,38 +20,10 @@ import {EXTENSION_ID} from "../../../constants";
  */
 export class ExtensionUtils {
     /**
-     *  This method calculate the time elapsed between two number representation of dates expressed as
-     *  number of milliseconds elapsed since January 1, 1970 00:00:00 UTC.
-     * @param startTime numeric represetation of a starting time
-     * @param endTime numeric representation of a ending time
-     */
-    public static calculateTimeElapsed(startTime: number, endTime: number): number | undefined {
-        return ((startTime && endTime) && (endTime >= startTime)) ? ((endTime - startTime) / 1000) : undefined;
-    }
-
-    /**
      * This method return the extension version declared into the package.json
      */
     public static getPackageVersion(): string {
         return vscode.extensions.getExtension(EXTENSION_ID).packageJSON.version;
-    }
-
-    // /**
-    //  * This method return the value of the instrumentation key necessary to create the telemetry reporter from an
-    //  * external file configuration. If the file doesn't exists it returns a generic value that will not be valid
-    //  * for collect telemetry event.
-    //  */
-    // public static getTelemetryKeyId(): string {
-    //     return fs.existsSync(this.getTelemetryResourcePath()) ? ExtensionUtils.getInstrumentationKey() : TELEMETRY_DEFAULT_CONTENT;
-    // }
-
-    // TODO: move to telemetry service
-    /**
-     * This method allows to anonymize the username value present in stack trace
-     * @param content stack trace
-     */
-    public static anonymizeContent(content: string): string {
-        return content.replace(new RegExp("\\" + sep + this.getUsername(), "g"), sep + "anonymous");
     }
 
     /**
@@ -63,24 +33,10 @@ export class ExtensionUtils {
         return (vscode.env) ? vscode.env.appName : "N.D.";
     }
 
-    // private static getTelemetryResourcePath() {
-    //     return vscode.Uri.file(
-    //         path.join(this.getExtensionPath(), "resources", "TELEMETRY_KEY")).fsPath;
-    // }
-
-    // TODO: update javadoc
     /**
-     *
+     * Return the physical location of the extension
      */
     public static getExtensionPath(): string {
         return vscode.extensions.getExtension(EXTENSION_ID).extensionPath;
-    }
-
-    // private static getInstrumentationKey(): string {
-    //     return Buffer.from(fs.readFileSync(ExtensionUtils.getTelemetryResourcePath(), "utf8").replace(/(\r\n|\n|\r)/gm, ""), "base64").toString();
-    // }
-
-    private static getUsername(): string {
-        return userInfo().username;
     }
 }
